+++ avl/_core/transaction.py
@@ -6,18 +6,15 @@
 from __future__ import annotations
 
 from collections.abc import Callable
-from typing import TYPE_CHECKING, Any
+from typing import Any
 
 from cocotb.triggers import Event
 from cocotb.utils import get_sim_time
 
 from .object import Object
 
-if TYPE_CHECKING:
-    from .component import Component
-
 class Transaction(Object):
-    def __init__(self, name: str, parent: Component) -> None:
+    def __init__(self, name: str, parent: Object|None) -> None:
         """
         Initialize a new Transaction.